+++ src/app/groups/[id]/management/members/_components/MemberCard.tsx
@@ -1,34 +1,5 @@
 import Image from 'next/image';
-import { Menu } from '@components/icons';
-
-interface MemberCardBtnsProps {
-  isLeader: boolean;
-  mode: 'curMembers' | 'waiting';
-}
-const MemberCardLeaderBtns = ({ isLeader, mode }: MemberCardBtnsProps) => {
-  return (
-    <>
-      {mode === 'curMembers' ? (
-        isLeader ? (
-          <span className="text-primary">대표</span>
-        ) : (
-          <Menu />
-        )
-      ) : (
-        <div>
-          <button type="button">수락</button>
-          <button type="button" className="ml-4 px-2.5 py-[0.438rem] bg-[#3B82F6] rounded-lg text-white font-semibold">
-            거절
-          </button>
-        </div>
-      )}
-    </>
-  );
-};
-
-const MemberCardStandardBtns = ({ isLeader, mode }: MemberCardBtnsProps) => {
-  return <>{mode === 'curMembers' && isLeader && <span className="text-primary">대표</span>}</>;
-};
+import MemberCardLeaderBtns from './MemberCardLeaderBtns';
 
 interface MemberCardProps {
   isLeader: boolean;
@@ -56,7 +27,7 @@ const MemberCard = ({ isLeader, isMe = false, mode }: MemberCardProps) => {
         {isLeaderUser ? (
           <MemberCardLeaderBtns isLeader={isLeader} mode={mode} />
         ) : (
-          <MemberCardStandardBtns isLeader={isLeader} mode={mode} />
+          <>{mode === 'curMembers' && isLeader && <span className="text-primary">대표</span>}</>
         )}
       </div>
     </div>
