+++ src/app/groups/[id]/page.tsx
@@ -1,9 +1,8 @@
-import GroupCard from '@components/common/GroupCard';
+import GroupCard from '@components/common/groupCard/GroupCard';
 import Header from '@components/common/Header';
 import TabPages from './_components/TabPages';
 
 const GroupPage = () => {
-
   return (
     <div className="w-full">
       {/* 헤더 영역 */}
@@ -15,7 +14,7 @@ const GroupPage = () => {
       <div className="w-full px-5 -mt-[2.75rem] z-10">
         <GroupCard
           groupInfo={{
-            created_at:'',
+            created_at: '',
             id: '123',
             name: '모각코!',
             description: '안녕하세요안녕하세요안녕하세요안녕하세요안녕하세요안녕하세요',
